+++ src/utils.rs
@@ -1,6 +1,6 @@
-pub const EPSILON: f32 = 0.00001;
+pub const EPSILON: f64 = 0.00001;
 
-pub fn equal(a: f32, b: f32) -> bool {
+pub fn equal(a: f64, b: f64) -> bool {
   if (a - b).abs() < EPSILON { return true; }
 
   false
